Fixed Notification Sending Issue.

+++ src/main/java/net/dahliasolutions/data/MessageRepository.java
@@ -3,6 +3,8 @@ package net.dahliasolutions.data;
 import net.dahliasolutions.models.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageRepository extends JpaRepository<Message, Long> {
+import java.math.BigInteger;
+
+public interface MessageRepository extends JpaRepository<Message, BigInteger> {
 
 }
