+++ src/main/java/net/dahliasolutions/services/MessageServiceInterface.java
@@ -3,11 +3,12 @@ package net.dahliasolutions.services;
 import net.dahliasolutions.models.Message;
 import net.dahliasolutions.models.Notification;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 public interface MessageServiceInterface {
 
-    Message createMessage(Long id, Notification notification);
+    Message createMessage(BigInteger id, Notification notification);
     void save(Message message);
-    Optional<Message> findById(Long id);
+    Optional<Message> findById(BigInteger id);
 }
